#279 Split test cases to reduce the maximum number of methods in EntityPersistenceNamingStrategyTest
